+++ common/src/main/java/dev/xylonity/tooltipoverhaul/client/layer/impl/PreviewBackgroundLayer.java
@@ -3,6 +3,7 @@ package dev.xylonity.tooltipoverhaul.client.layer.impl;
 import dev.xylonity.tooltipoverhaul.client.layer.ITooltipLayer;
 import dev.xylonity.tooltipoverhaul.client.layer.LayerDepth;
 import dev.xylonity.tooltipoverhaul.client.render.TooltipContext;
+import dev.xylonity.tooltipoverhaul.client.util.PositionUtils;
 import dev.xylonity.tooltipoverhaul.client.util.RenderUtils;
 import dev.xylonity.tooltipoverhaul.client.util.TextAxis;
 import dev.xylonity.tooltipoverhaul.config.TooltipsConfig;
@@ -31,14 +32,17 @@ public interface PreviewBackgroundLayer extends ITooltipLayer {
             Vec2 originalPos;
             Vec2 finalPos;
 
+            int extraX = PositionUtils.getSecondPanelPosition(context, TextAxis.X);
+            int extraY = PositionUtils.getSecondPanelPosition(context, TextAxis.Y);
+
             // Reposition the preview panel components if there is insufficient space at the left
             if (x1 < 0 && TooltipsConfig.AUTO_REPOSITION_PREVIEW_PANEL) {
-                originalPos = new Vec2(x0 + tooltipSizeX + margin * 2 + context.getPaddingX() + sizeX, y0);
-                finalPos = new Vec2(x1 + tooltipSizeX + margin * 2 + context.getPaddingX() + sizeX, y1);
+                originalPos = new Vec2(x0 + tooltipSizeX + margin * 2 + context.getPaddingX() + sizeX - extraX, y0 + extraY);
+                finalPos = new Vec2(x1 + tooltipSizeX + margin * 2 + context.getPaddingX() + sizeX - extraX, y1 + extraY);
             }
             else {
-                originalPos = new Vec2(x0, y0);
-                finalPos = new Vec2(x1, y1);
+                originalPos = new Vec2(x0 + extraX, y0 + extraY);
+                finalPos = new Vec2(x1 + extraX, y1 + extraY);
             }
 
             render(context, originalPos, finalPos);